Clean up split of responsibilities between stat.cpp/logger.cpp/screen.cpp
