+++ src/screen.cpp
@@ -17,33 +17,23 @@
  *           From Hewlett Packard Company.
  */
 
-/****
- * Screen.cpp : Simple curses & logfile encapsulation
- */
-
-
-#include <stdarg.h>
 #include <curses.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <signal.h>
-#include <errno.h>
-#include <sys/time.h>
-#include <sys/resource.h>
-#include <unistd.h>
 
 #include "screen.hpp"
-#include "screen_printer.hpp"
 #include "sipp.hpp"
 
 /* Export these so others needn't include curses.h */
 int key_backspace = KEY_BACKSPACE;
 int key_dc = KEY_DC;
 
-unsigned long screen_errors;
 int           screen_inited = 0;
-char          screen_exename[255];
+
+ScreenPrinter* sp;
+
+#ifdef RTP_STREAM
+double last_rtpstream_rate_out = 0;
+double last_rtpstream_rate_in = 0;
+#endif
 
 /* ERR is actually -1, but this prevents us from needing to use curses.h in
  * sipp.cpp. */
@@ -68,47 +58,9 @@ void screen_exit()
     screen_inited = 0;
 }
 
-void screen_show_errors() {
-    if (!screen_errors) {
-        return;
-    }
-
-    fprintf(stderr, "%s\n", screen_last_error);
-    if (screen_errors > 1) {
-        if (screen_logfile[0] != '\0') {
-            fprintf(stderr,
-                    "%s: There were more errors, see '%s' file\n",
-                    screen_exename, screen_logfile);
-        } else {
-            fprintf(stderr,
-                    "%s: There were more errors, enable -trace_err to log them.\n",
-                    screen_exename);
-        }
-    }
-    fflush(stderr);
-}
-
-void screen_clear()
-{
-    if (!screen_inited) {
-        return;
-    }
-
-    clear();
-    refresh();
-}
-
-void screen_set_exename(const char* exe_name)
-{
-    strncpy(screen_exename, exe_name, sizeof(screen_exename) - 1);
-}
-
 void screen_init()
 {
-    if (backgroundMode) {
-        return;
-    }
-    if (screen_inited) {
+    if (backgroundMode || screen_inited) {
         return;
     }
 
@@ -121,108 +73,829 @@ void screen_init()
     clear();
 }
 
-static void _screen_error(int fatal, bool use_errno, int error, const char *fmt, va_list ap)
+void print_statistics(int last)
 {
-    static unsigned long long count = 0;
-    char *c = screen_last_error;
-    struct timeval currentTime;
-
-    CStat::globalStat(fatal ? CStat::E_FATAL_ERRORS : CStat::E_WARNING);
-
-    GET_TIME (&currentTime);
-
-    c+= sprintf(c, "%s: ", CStat::formatTime(&currentTime));
-    c+= vsprintf(c, fmt, ap);
-    if (use_errno) {
-        c += sprintf(c, ", errno = %d (%s)", error, strerror(error));
+    if (backgroundMode == false && display_scenario) {
+      sp->redraw();
     }
-    screen_errors++;
+}
 
-    if (!error_lfi.fptr && print_all_responses) {
-        rotate_errorf();
-        if (error_lfi.fptr) {
-            fprintf(error_lfi.fptr, "%s: The following events occurred:\n",
-                    screen_exename);
-            fflush(error_lfi.fptr);
+void ScreenPrinter::print_closing_stats() {
+    M_last = true;
+    get_lines();
+    for (auto line : lines) {
+        printf("%s\n", line.c_str());
+    }
+
+    if (currentScreenToDisplay != DISPLAY_STAT_SCREEN) {
+        currentScreenToDisplay = DISPLAY_STAT_SCREEN;
+        get_lines();
+        for (auto line : lines) {
+            printf("%s\n", line.c_str());
+        }
+    }
+
+}
+
+void ScreenPrinter::print_to_file(FILE* f)
+{
+    get_lines();
+    for (auto line : lines) {
+        fprintf(f, "%s\n", line.c_str());
+    }
+}
+
+extern int command_mode;
+extern char* command_buffer;
+
+
+void ScreenPrinter::redraw()
+{
+    if (!M_headless) {
+        get_lines();
+        clear();
+        for (auto line : lines) {
+            printw("%s\n", line.c_str());
+        }
+
+        if (command_mode) {
+            printw("\nCommand: %s", command_buffer ? command_buffer : "");
+        }
+
+        refresh();
+    }
+}
+
+void ScreenPrinter::get_lines()
+{
+    lines.clear();
+    switch (currentScreenToDisplay) {
+    case DISPLAY_STAT_SCREEN:
+        lines.push_back("----------------------------- Statistics Screen "
+                        "------- [1-9]: Change Screen --");
+        draw_stats_screen();
+        break;
+    case DISPLAY_REPARTITION_SCREEN:
+        lines.push_back("---------------------------- Repartition Screen "
+                        "------- [1-9]: Change Screen --");
+        draw_repartition_screen(1);
+        break;
+    case DISPLAY_VARIABLE_SCREEN:
+        lines.push_back("----------------------------- Variables Screen "
+                        "-------- [1-9]: Change Screen --");
+        draw_vars_screen();
+        break;
+    case DISPLAY_TDM_MAP_SCREEN:
+        lines.push_back("------------------------------ TDM map Screen "
+                        "--------- [1-9]: Change Screen --");
+        draw_tdm_screen();
+        break;
+    case DISPLAY_SECONDARY_REPARTITION_SCREEN:
+        lines.push_back("--------------------------- Repartition " +
+                        std::to_string(currentRepartitionToDisplay) +
+                        " Screen ------ [1-9]: Change Screen --");
+        draw_repartition_screen(currentRepartitionToDisplay);
+        break;
+    case DISPLAY_SCENARIO_SCREEN:
+    default:
+        lines.push_back("------------------------------ Scenario Screen "
+                        "-------- [1-9]: Change Screen --");
+        draw_scenario_screen();
+        break;
+    }
+
+    char buf[80];
+    if (!M_last && screen_last_error[0]) {
+      char* errstart = screen_last_error;
+      int colonsleft = 3; /* We want to skip the time. */
+      while (*errstart && colonsleft) {
+        if (*errstart == ':') {
+          colonsleft--;
+        }
+        errstart++;
+      }
+      while (isspace(*errstart)) {
+        errstart++;
+      }
+      if (strlen(errstart) > 60) {
+        snprintf(buf, 80, "Last Error: %.60s...", errstart);
+      } else {
+        snprintf(buf, 80, "Last Error: %s", errstart);
+      }
+      lines.push_back(buf);
+    }
+
+    if (M_last) {
+        lines.push_back("------------------------------ Test Terminated "
+                   "--------------------------------");
+    } else if (quitting) {
+        lines.push_back(
+            "------- Waiting for active calls to end. Press [q] again "
+            "to force exit. -------");
+    } else if (paused) {
+        lines.push_back("----------------- Traffic Paused - Press [p] again to "
+                        "resume ------------------");
+    } else if (cpu_max) {
+        lines.push_back("-------------------------------- CPU CONGESTED "
+                        "---------------------------------");
+    } else if (outbound_congestion) {
+        lines.push_back("------------------------------ OUTBOUND CONGESTION "
+                        "-----------------------------");
+    } else {
+        if (creationMode == MODE_CLIENT) {
+            switch (thirdPartyMode) {
+            case MODE_MASTER:
+                lines.push_back(
+                    "-----------------------3PCC extended mode - Master "
+                    "side -------------------------");
+                break;
+            case MODE_3PCC_CONTROLLER_A:
+                lines.push_back(
+                    "----------------------- 3PCC Mode - Controller A "
+                    "side -------------------------");
+                break;
+            case MODE_3PCC_NONE:
+                lines.push_back(
+                    "------ [+|-|*|/]: Adjust rate ---- [q]: Soft exit "
+                    "---- [p]: Pause traffic -----");
+                break;
+            default:
+                ERROR("Internal error: creationMode=%d, thirdPartyMode=%d",
+                      creationMode, thirdPartyMode);
+            }
         } else {
-            if (screen_inited) {
-                sprintf(c, "%s: Unable to create '%s': %s.\n",
-                        screen_exename, screen_logfile, strerror(errno));
-            }
-            sipp_exit(EXIT_FATAL_ERROR);
-        }
-    }
-
-    if (error_lfi.fptr) {
-        count += fprintf(error_lfi.fptr, "%s", screen_last_error);
-        fflush(error_lfi.fptr);
-        if (ringbuffer_size && count > ringbuffer_size) {
-            rotate_errorf();
-            count = 0;
-        }
-        if (max_log_size && count > max_log_size) {
-            print_all_responses = 0;
-            if (error_lfi.fptr) {
-                fflush(error_lfi.fptr);
-                fclose(error_lfi.fptr);
-                error_lfi.fptr = NULL;
-                error_lfi.overwrite = false;
-            }
-        }
-    } else if (fatal) {
-        fprintf(stderr, "%s\n", screen_last_error);
-        fflush(stderr);
-    }
-
-    if (fatal) {
-        if (!screen_inited) {
-            if (error == EADDRINUSE) {
-                exit(EXIT_BIND_ERROR);
-            } else {
-                exit(EXIT_FATAL_ERROR);
-            }
-        } else {
-            if (error == EADDRINUSE) {
-                sipp_exit(EXIT_BIND_ERROR);
-            } else {
-                sipp_exit(EXIT_FATAL_ERROR);
+            assert(creationMode == MODE_SERVER);
+            switch (thirdPartyMode) {
+            case MODE_3PCC_A_PASSIVE:
+                lines.push_back(
+                    "------------------ 3PCC Mode - Controller A side "
+                    "(passive) --------------------");
+                break;
+            case MODE_3PCC_CONTROLLER_B:
+                lines.push_back(
+                    "----------------------- 3PCC Mode - Controller B "
+                    "side -------------------------");
+                break;
+            case MODE_MASTER_PASSIVE:
+                lines.push_back(
+                    "------------------ 3PCC extended mode - Master "
+                    "side (passive) --------------------");
+                break;
+            case MODE_SLAVE:
+                lines.push_back(
+                    "----------------------- 3PCC extended mode - Slave "
+                    "side -------------------------");
+                break;
+            case MODE_3PCC_NONE:
+                lines.push_back(
+                    "------------------------------ SIPp Server Mode "
+                    "-------------------------------");
+                break;
+            default:
+                ERROR("Internal error: creationMode=%d, thirdPartyMode=%d",
+                      creationMode, thirdPartyMode);
             }
         }
     }
 }
 
-extern "C" {
-    void ERROR(const char *fmt, ...)
-    {
-        va_list ap;
-        va_start(ap, fmt);
-        _screen_error(true, false, 0, fmt, ap);
-        va_end(ap);
-        assert(0);
+bool do_hide = true;
+bool show_index = false;
+
+void ScreenPrinter::draw_scenario_screen()
+{
+    char buf[80];
+    char left_buf[40];
+    char right_buf[80];
+    int divisor;
+    extern int pollnfds;
+
+    unsigned long long total_calls =
+        display_scenario->stats->GetStat(CStat::CPT_C_IncomingCallCreated) +
+        display_scenario->stats->GetStat(CStat::CPT_C_OutgoingCallCreated);
+    if (creationMode == MODE_SERVER) {
+        lines.push_back("  Port   Total-time  Total-calls  Transport");
+        snprintf(buf, 256, "  %-5d %6lu.%02lu s     %8llu  %s", local_port,
+                 clock_tick / 1000, (clock_tick % 1000) / 10, total_calls,
+                 TRANSPORT_TO_STRING(transport));
+        lines.push_back(buf);
+    } else {
+        assert(creationMode == MODE_CLIENT);
+        if (users >= 0) {
+            lines.push_back("     Users (length)   Port   Total-time  "
+                            "Total-calls  Remote-host");
+            snprintf(buf, 256,
+                     "%d (%d ms)   %-5d %6lu.%02lu s     %8llu  %s:%d(%s)",
+                     users, duration, local_port, clock_tick / 1000,
+                     (clock_tick % 1000) / 10, total_calls, remote_ip,
+                     remote_port, TRANSPORT_TO_STRING(transport));
+            lines.push_back(buf);
+        } else {
+            lines.push_back("     Call rate (length)   Port   Total-time  "
+                            "Total-calls  Remote-host");
+            snprintf(
+                buf, 256,
+                "%3.1f(%d ms)/%5.3fs   %-5d %6lu.%02lu s     %8llu  %s:%d(%s)",
+                rate, duration, (double)rate_period_ms / 1000.0, local_port,
+                clock_tick / 1000, (clock_tick % 1000) / 10, total_calls,
+                remote_ip, remote_port, TRANSPORT_TO_STRING(transport));
+            lines.push_back(buf);
+        }
+    }
+    lines.push_back("");
+    /* 1st line */
+    unsigned long ms_since_last_tick = clock_tick - last_report_time;
+    if (total_calls < stop_after) {
+        snprintf(left_buf, 40, "%llu new calls during %lu.%03lu s period",
+                 display_scenario->stats->GetStat(
+                     CStat::CPT_PD_IncomingCallCreated) +
+                     display_scenario->stats->GetStat(
+                         CStat::CPT_PD_OutgoingCallCreated),
+                 ms_since_last_tick / 1000, ms_since_last_tick % 1000);
+    } else {
+        snprintf(left_buf, 40,
+                 "Call limit reached (-m %lu), %lu.%03lu s period ", stop_after,
+                 ms_since_last_tick / 1000, ms_since_last_tick % 1000);
+    }
+    snprintf(right_buf, 40, "%lu ms scheduler resolution",
+             ms_since_last_tick / std::max(scheduling_loops, 1ul));
+    snprintf(buf, 80, "  %-38s  %-40s", left_buf, right_buf);
+    lines.push_back(buf);
+
+    /* 2nd line */
+    if (creationMode == MODE_SERVER) {
+        snprintf(left_buf, 40, "%llu calls",
+                 display_scenario->stats->GetStat(CStat::CPT_C_CurrentCall));
+    } else {
+        snprintf(left_buf, 40, "%llu calls (limit %u)",
+                 display_scenario->stats->GetStat(CStat::CPT_C_CurrentCall),
+                 open_calls_allowed);
+    }
+    snprintf(
+        buf, 80, "  %-38s  Peak was %llu calls, after %llu s", left_buf,
+        display_scenario->stats->GetStat(CStat::CPT_C_CurrentCallPeak),
+        display_scenario->stats->GetStat(CStat::CPT_C_CurrentCallPeakTime));
+    lines.push_back(buf);
+
+    snprintf(buf, 80, "  %d Running, %d Paused, %d Woken up",
+             last_running_calls, last_paused_calls, last_woken_calls);
+    last_woken_calls = 0;
+    lines.push_back(buf);
+
+    /* 3rd line dead call msgs, and optional out-of-call msg */
+    snprintf(left_buf, 40, "%llu dead call msg (discarded)",
+             display_scenario->stats->GetStat(CStat::CPT_G_C_DeadCallMsgs));
+    if (creationMode == MODE_CLIENT) {
+        snprintf(
+            buf, 80, "  %-38s  %llu out-of-call msg (discarded)", left_buf,
+            display_scenario->stats->GetStat(CStat::CPT_G_C_OutOfCallMsgs));
+    } else {
+        snprintf(buf, 80, "  %-38s", left_buf);
+    }
+    lines.push_back(buf);
+
+    if (compression) {
+        snprintf(buf, 80, "  Comp resync: %d sent, %d recv", resynch_send,
+                 resynch_recv);
+        lines.push_back(buf);
     }
 
-    void ERROR_NO(const char *fmt, ...)
-    {
-        va_list ap;
-        va_start(ap, fmt);
-        _screen_error(true, true, errno, fmt, ap);
-        va_end(ap);
-        assert(0);
+    /* 4th line , sockets and optional errors */
+    snprintf(left_buf, 40, "%d open sockets", pollnfds);
+    snprintf(buf, 80, "  %-38s  %lu/%lu/%lu %s errors (send/recv/cong)",
+             left_buf, nb_net_send_errors, nb_net_recv_errors, nb_net_cong,
+             TRANSPORT_TO_STRING(transport));
+    lines.push_back(buf);
+
+#ifdef PCAPPLAY
+    /* if has media abilities */
+    if (hasMedia != 5) {
+        snprintf(left_buf, 40, "%lu Total RTP pckts sent ",
+                rtp_pckts_pcap);
+        if (ms_since_last_tick) {
+            snprintf(buf, 80, "  %-38s  %lu.%03lu last period RTP rate (kB/s)",
+                    left_buf,
+                    rtp_bytes_pcap / ms_since_last_tick,
+                    rtp_bytes_pcap % ms_since_last_tick);
+        }
+        rtp_bytes_pcap = 0;
+        rtp2_bytes_pcap = 0;
+        lines.push_back(buf);
+    }
+#endif
+#ifdef RTP_STREAM
+    /* if we have rtp stream thread running */
+    if (rtpstream_numthreads) {
+        unsigned long tempbytes;
+        unsigned long last_tick = clock_tick;
+        /* Saved clock_tick to last_tick and use that in calcs since clock tick */
+        /* can change during calculations.                                      */
+        if (ms_since_last_tick) {
+            tempbytes = rtpstream_bytes_out;
+            /* Calculate integer and fraction parts of rtp bandwidth; this value
+             * will be saved and reused in the case where last_tick==last_report_time
+             */
+            last_rtpstream_rate_out = ((double)tempbytes) / ms_since_last_tick;
+            /* Potential race condition betwen multiple threads updating the
+             * rtpstream_bytes value. We subtract the saved tempbytes value
+             * rather than setting it to zero to minimise the chances of missing
+             * an update to rtpstream_bytes [update between printing stats and
+             * zeroing the counter]. Ideally we would atomically subtract
+             * tempbytes from rtpstream_bytes.
+             */
+            rtpstream_bytes_out -= tempbytes;
+            tempbytes = rtpstream_bytes_in;
+            last_rtpstream_rate_in = ((double)tempbytes) / ms_since_last_tick;
+            rtpstream_bytes_in -= tempbytes;
+        }
+        snprintf(left_buf, 40, "%lu Total RTP pckts sent", rtpstream_pckts);
+        snprintf(buf, 80,"  %-38s  %.3f kB/s RTP OUT",
+                left_buf, last_rtpstream_rate_out);
+        lines.push_back(buf);
+
+        snprintf(left_buf, 40, "%lu RTP sending threads active", rtpstream_numthreads);
+        snprintf(buf, 80, "  %-38s  %.3f kB/s RTP IN",
+                 left_buf, last_rtpstream_rate_in);
+        lines.push_back(buf);
+    }
+#endif
+
+    /* 5th line, RTP echo statistics */
+    if (rtp_echo_enabled && media_socket_audio > 0) {
+        snprintf(left_buf, 40, "%lu Total echo RTP pckts 1st stream",
+                rtp_pckts);
+
+        if (ms_since_last_tick) {
+            snprintf(buf, 80, "  %-38s  %lu.%03lu last period RTP rate (kB/s)",
+                    left_buf,
+                    rtp_bytes / ms_since_last_tick,
+                    rtp_bytes % ms_since_last_tick);
+            lines.push_back(buf);
+        }
+
+        snprintf(left_buf, 40, "%lu Total echo RTP pckts 2nd stream",
+                rtp2_pckts);
+        if (ms_since_last_tick) {
+            snprintf(buf, 80, "  %-38s  %lu.%03lu last period RTP rate (kB/s)",
+                    left_buf,
+                    rtp2_bytes / ms_since_last_tick,
+                    rtp2_bytes % ms_since_last_tick);
+            lines.push_back(buf);
+        }
+        rtp_bytes = 0;
+        rtp2_bytes = 0;
     }
 
-    void WARNING(const char *fmt, ...)
-    {
-        va_list ap;
-        va_start(ap, fmt);
-        _screen_error(false, false, 0, fmt, ap);
-        va_end(ap);
+    /* Scenario counters */
+    lines.push_back("");
+    if (!lose_packets) {
+        snprintf(buf, 80,
+                 "                                 "
+                 "Messages  Retrans   Timeout   Unexpected-Msg");
+    } else {
+        snprintf(buf, 80,
+                 "                                 "
+                 "Messages  Retrans   Timeout   Unexp.    Lost");
+    }
+    lines.push_back(buf);
+
+    for (unsigned long index = 0; index < display_scenario->messages.size();
+         index++) {
+        buf[0] = 0;
+        message* curmsg = display_scenario->messages[index];
+
+        if (do_hide && curmsg->hide) {
+            continue;
+        }
+        int buf_len = 0;
+        if (show_index) {
+            buf_len += snprintf(buf + buf_len, 80 - buf_len, "%-2lu:", index);
+        }
+
+        if (SendingMessage* src = curmsg->send_scheme) {
+            if (creationMode == MODE_SERVER) {
+                if (src->isResponse()) {
+                    buf_len += snprintf(buf + buf_len, 80 - buf_len,
+                                        "  <---------- %-10d ", src->getCode());
+                } else {
+                    buf_len +=
+                        snprintf(buf + buf_len, 80 - buf_len,
+                                 "  <---------- %-10s ", src->getMethod());
+                }
+            } else {
+                if (src->isResponse()) {
+                    buf_len += snprintf(buf + buf_len, 80 - buf_len,
+                                        "  %10d ----------> ", src->getCode());
+                } else {
+                    buf_len +=
+                        snprintf(buf + buf_len, 80 - buf_len,
+                                 "  %10s ----------> ", src->getMethod());
+                }
+            }
+
+            if (curmsg->start_rtd) {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len, " B-RTD%d ",
+                                    curmsg->start_rtd);
+            } else if (curmsg->stop_rtd) {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len, " E-RTD%d ",
+                                    curmsg->stop_rtd);
+            } else {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len, "        ");
+            }
+
+            if (curmsg->retrans_delay) {
+                buf_len += snprintf(
+                    buf + buf_len, 80 - buf_len, "%-9lu %-9lu %-9lu %-9s %-9s",
+                    curmsg->nb_sent, curmsg->nb_sent_retrans,
+                    curmsg->nb_timeout, "" /* Unexpected */,
+                    (lose_packets && curmsg->nb_lost)
+                        ? std::to_string(curmsg->nb_lost).c_str()
+                        : "");
+            } else {
+                buf_len += snprintf(
+                    buf + buf_len, 80 - buf_len, "%-9lu %-9lu %-9s %-9s %-9s",
+                    curmsg->nb_sent, curmsg->nb_sent_retrans, "", /* Timeout. */
+                    "" /* Unexpected. */,
+                    (lose_packets && curmsg->nb_lost)
+                        ? std::to_string(curmsg->nb_lost).c_str()
+                        : "");
+            }
+        } else if (curmsg->recv_response) {
+            if (creationMode == MODE_SERVER) {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len,
+                                    "  ----------> %-10d ", curmsg->recv_response);
+            } else {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len,
+                                    "  %10d <---------- ", curmsg->recv_response);
+            }
+
+            if (curmsg->start_rtd) {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len, " B-RTD%d ",
+                                    curmsg->start_rtd);
+            } else if (curmsg->stop_rtd) {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len, " E-RTD%d ",
+                                    curmsg->stop_rtd);
+            } else {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len, "        ");
+            }
+
+            buf_len += snprintf(buf + buf_len, 80 - buf_len,
+                                "%-9ld %-9ld %-9ld %-9ld %-9s", curmsg->nb_recv,
+                                curmsg->nb_recv_retrans, curmsg->nb_timeout,
+                                curmsg->nb_unexp,
+                                (lose_packets && curmsg->nb_lost)
+                                    ? std::to_string(curmsg->nb_lost).c_str()
+                                    : "");
+        } else if (curmsg->pause_distribution ||
+                   (curmsg->pause_variable != -1)) {
+            char* desc = curmsg->pause_desc;
+            if (!desc) {
+                desc = (char*)malloc(24);
+                if (curmsg->pause_distribution) {
+                    desc[0] = '\0';
+                    curmsg->pause_distribution->timeDescr(desc, 23);
+                } else {
+                    snprintf(desc, 23, "$%s",
+                             display_scenario->allocVars->getName(
+                                 curmsg->pause_variable));
+                }
+                desc[23] = '\0';
+                curmsg->pause_desc = desc;
+            }
+            int len = strlen(desc) < 9 ? 9 : strlen(desc);
+
+            if (creationMode == MODE_SERVER) {
+                snprintf(left_buf, 40, "  [%9s] Pause%*s", desc,
+                        23 - len > 0 ? 23 - len : 0, "");
+            } else {
+                snprintf(left_buf, 40, "       Pause [%9s]%*s", desc,
+                        18 - len > 0 ? 18 - len : 0, "");
+            }
+
+            snprintf(buf, 80, "%s%-9d                     %-9lu",
+                     left_buf,
+                     curmsg->sessions,
+                     curmsg->nb_unexp);
+        } else if (curmsg->recv_request) {
+            if (creationMode == MODE_SERVER) {
+                buf_len +=
+                    snprintf(buf + buf_len, 80 - buf_len,
+                             "  ----------> %-10s ", curmsg->recv_request);
+            } else {
+                buf_len +=
+                    snprintf(buf + buf_len, 80 - buf_len, "  %10s <---------- ",
+                             curmsg->recv_request);
+            }
+
+            if (curmsg->start_rtd) {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len, " B-RTD%d ",
+                                    curmsg->start_rtd);
+            } else if (curmsg->stop_rtd) {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len, " E-RTD%d ",
+                                    curmsg->stop_rtd);
+            } else {
+                buf_len += snprintf(buf + buf_len, 80 - buf_len, "        ");
+            }
+
+            buf_len += snprintf(buf + buf_len, 80 - buf_len,
+                                "%-9ld %-9ld %-9ld %-9ld %-9s", curmsg->nb_recv,
+                                curmsg->nb_recv_retrans, curmsg->nb_timeout,
+                                curmsg->nb_unexp,
+                                (lose_packets && curmsg->nb_lost)
+                                    ? std::to_string(curmsg->nb_lost).c_str()
+                                    : "");
+        } else if (curmsg->M_type == MSG_TYPE_NOP) {
+            if (curmsg->display_str) {
+                snprintf(buf, 80, " %s", curmsg->display_str);
+            } else {
+                snprintf(buf, 80, "              [ NOP ]              ");
+            }
+        } else if (curmsg->M_type == MSG_TYPE_RECVCMD) {
+            snprintf(left_buf, 40, "    [ Received Command ]         ");
+            snprintf(buf, 80, "%s%-9ld %-9s %-9s %-9s",
+                     left_buf,
+                     curmsg->M_nbCmdRecv,
+                     "",
+                     curmsg->retrans_delay ? std::to_string(curmsg->nb_timeout).c_str() : "",
+                     "");
+        } else if (curmsg->M_type == MSG_TYPE_SENDCMD) {
+            snprintf(left_buf, 40, "        [ Sent Command ]         ");
+            snprintf(buf, 80, "%s%-9lu %-9s           %-9s",
+                     left_buf,
+                     curmsg->M_nbCmdSent,
+                     "",
+                     "");
+        } else {
+            ERROR("Scenario command not implemented in display");
+        }
+
+        lines.push_back(buf);
+        if (curmsg->crlf) {
+            lines.push_back("");
+        }
+    }
+}
+
+// Warning! All DISPLAY_ macros must be called where 'buf' and 'lines' are defined.
+#define DISPLAY_LINE() \
+    sprintf(buf," ------------------------------------------------------------------------------ "); lines.push_back(buf);
+#define DISPLAY_DLINE() \
+    sprintf(buf,"================================================================================"); lines.push_back(buf);
+#define DISPLAY_CROSS_LINE() \
+    sprintf(buf,"-------------------------+---------------------------+--------------------------"); lines.push_back(buf);
+
+#define DISPLAY_HEADER() \
+    sprintf(buf,"  Counter Name           | Periodic value            | Cumulative value"); lines.push_back(buf);
+#define DISPLAY_TXT_COL(T1, V1, V2) \
+    sprintf(buf,"  %-22.22s | %-25.25s | %-24.24s ", T1, V1, V2); lines.push_back(buf);
+#define DISPLAY_VAL_RATEF_COL(T1, V1, V2) \
+    sprintf(buf,"  %-22.22s | %8.3f cps              | %8.3f cps             ", T1, V1, V2); lines.push_back(buf);
+#define DISPLAY_2VAL(T1, V1, V2) \
+    sprintf(buf,"  %-22.22s | %8llu                  | %8llu                 ", T1, V1, V2); lines.push_back(buf);
+#define DISPLAY_CUMUL(T1, V1) \
+    sprintf(buf,"  %-22.22s |                           | %8llu                 ", T1, V1); lines.push_back(buf);
+#define DISPLAY_PERIO(T1, V1) \
+    sprintf(buf,"  %-22.22s | %8llu                  |                          ", T1, V1); lines.push_back(buf);
+#define DISPLAY_VALF(T1, V1) \
+    sprintf(buf,"  %-22.22s | %8.3f ms                                          ", T1, V1); lines.push_back(buf);
+#define DISPLAY_VAL_RATEF(T1, V1) \
+    sprintf(buf,"  %-22.22s | %8.3f cps                                         ", T1, V1); lines.push_back(buf);
+#define DISPLAY_VAL_RATE(T1, V1) \
+    sprintf(buf,"  %-22.22s | %8d cps                                         ", T1, V1); lines.push_back(buf);
+#define DISPLAY_VAL(T1, V1) \
+    sprintf(buf,"  %-22.22s : %8d                                             ", T1, V1); lines.push_back(buf);
+#define DISPLAY_2VALF(T1, V1, T2, V2) \
+    sprintf(buf,"  %-22.22s : %8.2f  | %-7.7s : %8.2f                       ", T1, V1, T2, V2); lines.push_back(buf);
+#define DISPLAY_3VAL(T1, V1, T2, V2, T3, V3) \
+    sprintf(buf,"  %-22.22s : %8d  | %-7.7s : %8d  | %-12.12s : %5d ", T1, V1, T2, V2, T3, V3); lines.push_back(buf);
+#define DISPLAY_3VALF(T1, V1, T2, V2, T3, V3) \
+    sprintf(buf,"  %-22.22s : %8.3f  | %-7.7s : %8.3f  | %-12.12s : %5.1f ", T1, V1, T2, V2, T3, V3); lines.push_back(buf);
+#define DISPLAY_TXT(T1, V1) \
+    sprintf(buf,"  %-22.22s | %-52.52s ", T1, V1); lines.push_back(buf);
+#define DISPLAY_INFO(T1) \
+    sprintf(buf,"  %-77.77s ", T1); lines.push_back(buf);
+#define DISPLAY_REPART(T1, T2, V1) \
+    sprintf(buf,"    %8d ms <= n <  %8d ms : %10lu  %-29.29s ", T1, T2, V1, ""); lines.push_back(buf);
+#define DISPLAY_LAST_REPART(T1, V1) \
+    sprintf(buf,"    %14.14s n >= %8d ms : %10lu  %-29.29s ", "", T1, V1, ""); lines.push_back(buf);
+
+void ScreenPrinter::draw_stats_screen()
+{
+    long   localElapsedTime, globalElapsedTime ;
+    struct timeval currentTime;
+    float  averageCallRate;
+    float  realInstantCallRate;
+    unsigned long numberOfCall;
+    CStat* s = display_scenario->stats;
+    char buf[256];
+
+    GET_TIME (&currentTime);
+    // computing the real call rate
+    globalElapsedTime   = s->computeDiffTimeInMs (&currentTime, &s->M_startTime);
+    localElapsedTime    = s->computeDiffTimeInMs (&currentTime, &s->M_pdStartTime);
+    // the call rate is for all the call : incoming and outgoing
+    numberOfCall        = (s->M_counters[s->CPT_C_IncomingCallCreated] +
+                           s->M_counters[s->CPT_C_OutgoingCallCreated]);
+    averageCallRate     = (globalElapsedTime > 0 ?
+                           1000*(float)numberOfCall/(float)globalElapsedTime :
+                           0.0);
+    numberOfCall        = (s->M_counters[s->CPT_PD_IncomingCallCreated] +
+                           s->M_counters[s->CPT_PD_OutgoingCallCreated]);
+    realInstantCallRate = (localElapsedTime  > 0 ?
+                           1000*(float)numberOfCall / (float)localElapsedTime :
+                           0.0);
+
+    // build and display header info
+    DISPLAY_TXT ("Start Time  ", s->formatTime(&s->M_startTime));
+    DISPLAY_TXT ("Last Reset Time", s->formatTime(&s->M_pdStartTime));
+    DISPLAY_TXT ("Current Time", s->formatTime(&currentTime));
+
+    // printing the header in the middle
+    DISPLAY_CROSS_LINE();
+    DISPLAY_HEADER();
+    DISPLAY_CROSS_LINE();
+
+    DISPLAY_TXT_COL ("Elapsed Time",
+                     s->msToHHMMSSus(localElapsedTime),
+                     s->msToHHMMSSus(globalElapsedTime));
+
+    DISPLAY_VAL_RATEF_COL ("Call Rate",  realInstantCallRate, averageCallRate);
+    DISPLAY_CROSS_LINE ();
+
+    DISPLAY_2VAL  ("Incoming calls created",
+                   s->M_counters[s->CPT_PD_IncomingCallCreated],
+                   s->M_counters[s->CPT_C_IncomingCallCreated]);
+    DISPLAY_2VAL  ("Outgoing calls created",
+                   s->M_counters[s->CPT_PD_OutgoingCallCreated],
+                   s->M_counters[s->CPT_C_OutgoingCallCreated]);
+    DISPLAY_CUMUL ("Total Calls created", s->M_counters[s->CPT_C_IncomingCallCreated] +
+                   s->M_counters[s->CPT_C_OutgoingCallCreated]);
+    DISPLAY_PERIO ("Current Calls",
+                   s->M_counters[s->CPT_C_CurrentCall]);
+
+    if (s->M_genericMap.size()) {
+        DISPLAY_CROSS_LINE ();
+    }
+    for (unsigned int i = 1; i < s->M_genericMap.size() + 1; i++) {
+        char *disp = (char *)malloc(20 + strlen(s->M_genericDisplay[i]));
+        sprintf(disp, "Counter %s", s->M_genericDisplay[i]);
+
+        DISPLAY_2VAL(disp, s->M_genericCounters[(i - 1)* GENERIC_TYPES + GENERIC_PD], s->M_genericCounters[(i - 1) * GENERIC_TYPES + GENERIC_C]);
+        free(disp);
     }
 
-    void WARNING_NO(const char *fmt, ...)
+    DISPLAY_CROSS_LINE ();
+    DISPLAY_2VAL  ("Successful call",
+                   s->M_counters[s->CPT_PD_SuccessfulCall],
+                   s->M_counters[s->CPT_C_SuccessfulCall]);
+    DISPLAY_2VAL  ("Failed call",
+                   s->M_counters[s->CPT_PD_FailedCall],
+                   s->M_counters[s->CPT_C_FailedCall]);
+
+    DISPLAY_CROSS_LINE ();
+    for (int i = 1; i <= s->nRtds(); i++) {
+        char buf2[80];
+
+        sprintf(buf2, "Response Time %s", s->M_revRtdMap[i]);
+        DISPLAY_TXT_COL (buf2,
+                         s->msToHHMMSSus( (unsigned long)s->computeRtdMean(i, GENERIC_PD)),
+                         s->msToHHMMSSus( (unsigned long)s->computeRtdMean(i, GENERIC_C)));
+    }
+    DISPLAY_TXT_COL ("Call Length",
+                     s->msToHHMMSSus( (unsigned long)s->computeMean(s->CPT_PD_AverageCallLength_Sum, s->CPT_PD_NbOfCallUsedForAverageCallLength ) ),
+                     s->msToHHMMSSus( (unsigned long)s->computeMean(s->CPT_C_AverageCallLength_Sum, s->CPT_C_NbOfCallUsedForAverageCallLength) ));
+}
+
+void ScreenPrinter::draw_repartition_screen(int which)
+{
+    char buf[80];
+    char buf2[80];
+    CStat* s = display_scenario->stats;
+    if (which > s->nRtds()) {
+        DISPLAY_INFO ("  <No repartion defined>");
+        return;
+    }
+
+    snprintf(buf2, 80, "Average Response Time Repartition %s", s->M_revRtdMap[which]);
+    DISPLAY_INFO(buf2);
+    draw_repartition_detailed(s->M_ResponseTimeRepartition[which - 1],
+                              s->M_SizeOfResponseTimeRepartition);
+
+    if (which == 1)
     {
-        va_list ap;
-        va_start(ap, fmt);
-        _screen_error(false, true, errno, fmt, ap);
-        va_end(ap);
+        // Primary repartition screen
+        DISPLAY_INFO("Average Call Length Repartition");
+        draw_repartition_detailed(s->M_CallLengthRepartition,
+                                  s->M_SizeOfCallLengthRepartition);
+    }
+}
+
+void ScreenPrinter::draw_repartition_detailed(CStat::T_dynamicalRepartition * tabRepartition,
+                                            int sizeOfTab)
+{
+    char buf[80];
+    if(tabRepartition != NULL) {
+        for(int i=0; i<(sizeOfTab-1); i++) {
+            if(i==0) {
+                DISPLAY_REPART(0, tabRepartition[i].borderMax,
+                               tabRepartition[i].nbInThisBorder);
+            } else {
+                DISPLAY_REPART(tabRepartition[i-1].borderMax,
+                               tabRepartition[i].borderMax,
+                               tabRepartition[i].nbInThisBorder);
+            }
+        }
+        DISPLAY_LAST_REPART (tabRepartition[sizeOfTab-1].borderMax,
+                             tabRepartition[sizeOfTab-1].nbInThisBorder);
+    } else {
+        DISPLAY_INFO ("  <No repartion defined>");
+    }
+}
+
+void ScreenPrinter::draw_vars_screen()
+{
+    CActions* actions;
+    CAction* action;
+    bool found;
+    char buf[80];
+
+    lines.push_back("Action defined Per Message :");
+    found = false;
+    for (unsigned int i = 0; i < display_scenario->messages.size(); i++) {
+        message* curmsg = display_scenario->messages[i];
+        actions = curmsg->M_actions;
+        if (actions != NULL) {
+            switch (curmsg->M_type) {
+            case MSG_TYPE_RECV:
+                snprintf(buf, 80, "=> Message[%u] (Receive Message) - "
+                         "[%d] action(s) defined :",
+                         i, actions->getActionSize());
+                break;
+            case MSG_TYPE_RECVCMD:
+                snprintf(buf, 80, "=> Message[%u] (Receive Command Message) - "
+                         "[%d] action(s) defined :",
+                         i, actions->getActionSize());
+                break;
+            default:
+                snprintf(buf, 80, "=> Message[%u] - [%d] action(s) defined :", i,
+                       actions->getActionSize());
+                break;
+            }
+            lines.push_back(buf);
+
+            for (int j = 0; j < actions->getActionSize(); j++) {
+                action = actions->getAction(j);
+                if (action != NULL) {
+                    int printed = snprintf(buf, 80, "   --> action[%d] = ", j);
+                    action->printInfo(buf + printed, 80 - printed);
+                    lines.push_back(buf);
+                    found = true;
+                }
+            }
+        }
+    }
+    if (!found) {
+        lines.push_back("=> No action found on any messages");
+    }
+
+    lines.push_back("");
+    for (unsigned int i = 0;
+         i < (display_scenario->messages.size() + 6 - lines.size()); i++) {
+      lines.push_back("");
+    }
+}
+
+void ScreenPrinter::draw_tdm_screen()
+{
+    int i = 0;
+    int in_use = 0;
+    int height = (tdm_map_a + 1) * (tdm_map_b + 1);
+    int width = (tdm_map_c + 1);
+    int total_circuits = height * width;
+    char buf[80] = {0};
+
+    lines.push_back("TDM Circuits in use:");
+    while (i < total_circuits) {
+        int buf_position = std::min(79, i % width);
+        if (tdm_map[i]) {
+            buf[buf_position] = '*';
+            in_use++;
+        } else {
+            buf[buf_position] = '.';
+        }
+        i++;
+        if (buf_position == (width - 1)) {
+            lines.push_back(buf);
+            memset(buf, 0, 80);
+        }
+    }
+    lines.push_back("");
+    snprintf(buf, 80, "%d/%d circuits (%d%%) in use", in_use, total_circuits,
+           int(100 * in_use / total_circuits));
+    lines.push_back(buf);
+    for (unsigned int i = 0;
+         i < (display_scenario->messages.size() + 8 - height);
+         i++) {
+        lines.push_back("");
     }
 }
